build: first pass at builder interface
A builder takes a path on the FS, runs a cmd in them, and returns a new
path on the file system with those commands run, the output of those cmds,
and an error.

There could be different builder implementations: one for building locally,
one for remotely, one that just makes new directories, one that makes new
layers in overlayFS, etc.