image: history shouldn't grow exponentially, let's at least aim for linearly
Prior to this patch when we loaded the existing history from disk
(in `NewImageHistory`) we would _re-persist_ all of the images that had
already been persisted.

To prevent this, I broke out `history.Add` in to two functions:
1. `history.Load` now adds an image to the in memory store
2. `history.AddAndPersist` loads an image and persists it to disk

We only call `history.Load` in `NewImageHistory` to prevent the exponential
hsitory problem.