engine: simple state machine for docker compose events (#959)
Docker Compose container events are very basic, and undocumented. From
my experiments here's what I've deduced that they mean:

* die - a process has died
* kill - a SIGKILL was sent to a container
* stop - a container has stopped
* create - a request for a container to be created was received
* start - a container has started running
* restart - a container has succesfully restarted

Naively to detect if a process has crashed would just check for "die".
This is insufficient. "die" can sometimes occur subsequent to a "kill"
message. For example the full event stream for a `docker restart` is:
* die
* start
* restart

So we have to be a little smarter. In English, a docker compose process
has crashed if it emits a "die" message but it isn't in the middle of
starting or restarting. To achieve that I wrote a little state machine
in the `handleDockerComposeEventAction`.