integration: fix exec test (#1625)
Prior to this change the exec test wasn't really testing exec. When
we tried to do an exec update it would fail due to incompatibilites
between GNU tar and busybox tar. This would cause the update to fall
back to an image build, which caused the test to pass.

Let's make sure that the pods dont change AND install GNU tar in the
container so that the test passes.

While this feels a little gross, it's at least testing more than it was.