hud/server: create snapshot HTTP bootstrap simulacrum
With this change someone working on Tilt locally can navigate to
http://localhost:10350/snapshot/aaaaaa and Tilt will make a request to
localhost:10350/api/snapshot/aaaaaa which will serialize the current
Tilt engine state in to the snapshot format that the frontend understands
and return it.

Tilt will bootstrap off of this HTTP request thus simulating snapshots.

This addresses only part of ch3234