docker: use new WithAPIVersionNegotiation Opt to actually negotiate API version w/ server instead of erroring
in the old code, we applied `NegotiateAPIVersion` as an opt, i.e. a
function that we ran the `Client` through when we were creating it.
This resulted in us trying to negotiate with the server--which involved
pinging the server--before we had set a scheme on the `Client` (which
happens later in the client creation process). This, the API negotiation
always failed silently, and we would set the lowest possible version on
the client.