webview: simple GRPC server
The point of this commit is just to validate the skeleton of the GRPC server,
not to have something useful.

This will be expanded in the future to fully implement all of the APIs
in the HUD webserver (analytics, triggers, snapshots, etc) EXCEPT the
websocket.

Other future work includes generating a TypeScript client for use
in the web frontend.