model: store logs by segment instead of by line
A key API we want to be able to support is "give me all the logs since checkpoint X".
Storing the logs in an append-only data structure makes this much easier to support,
even if it makes it a little bit harder to print the logs line-by-line.

Storing the logs line-by-line meant we had to do backtracking and make things mutable.