tiltfile: error in default_registry if registry is not canonical
This causes a stack trace to be printed during Tiltfile execution
which makes it more obvious where this error is coming from.

Previously the error would only surface when the default registry
was being transformed in to the engine state model, at which point
we don't know what line it occurred on.

Fixes #2768