analytics: send anonymized analytics by default. update the README and nudge to reflect this
We're having trouble understanding the early part of the Tilt experience.
Are people making it to the web UI?
Are they able to run a project successfully?
Are they able to set up a local kubernetes cluster?

We're going to try enabling it by default on tilt v0.12 and see
what happens. Previous PRs have cleaned up some of the existing
analytics behavior so that it doesn't pick up CI and respects DO_NOT_TRACK=1,
so we feel more comfortable with this.