store: Simplify subscription notification
In the old world, every Notify() would create a new goroutine.
The goroutine would block until the subscriber was done processing the last event,
then double-check if it still needed to process.
This made debugging difficult, because you could have 1000s of
goroutines blocked on a single slow subscriber.

In the new world, there are at most two goroutines in flight.