store: clear pending files even on error builds
This is a small code change, but should make the possible build control states
much easier to reason about and display.

This logic is vetigial from the days when we had fast_build and incremental
image builds. If your image build failed, and you wanted to retry it, then Tilt
would need to sync both the recently changed files and the files from the
*failed* build.

This can't really happen anymore where you have partial failed syncs.
If the sync step fails, we'll fall back to an image build.
If the run step fails, we don't need to re-sync the same files.