remove auto-namespace creation. Fixes  https://github.com/tilt-dev/tilt/issues/3349
helm3 has moved away from this behavior, and in retrospect, it seems
to create more problems than it solves

As a work-around, people should create a new namespace
themselves if that's what they want.

This is basically a revert of https://github.com/tilt-dev/tilt/pull/2418

I also considered adding some sort of EnsureNamespace
primitive in K8sTarget, which might be more generally useful,
but was worried it added a lot of complexity and made this
harder to implement as a tilt extension.