engine: remove LastSuccessfulResult
The fundamental problem is that, by definition, all concrete BuildResult objects
represent success. If an image build fails, we return a nil BuildResult.

The notion of "BuildResult" gets slippery when we talk about live update.
If the in-container RUN step failed, the live-updater returns a valid BuildResult
(containing the container id) AND an error. In some contexts, we want to
treat that as a success (because the container is still live and can get future
live updates), but in other cases (like error reporting), we want to report
this as an error.

Better to just get rid of the concept of "SuccessfulResult" completely,
and be more precise about what we want.