engine: fix bugs in image build caching
this is the source of a lot of the infinite rebuild problems in
https://github.com/tilt-dev/tilt/pull/3362

There are a lot of weird cases where the original image existence check would
fail, and so we would deliberately rebuild the image. This makes the check more
robust, and adds logging when it does fail so that we have some hope of figuring
out the problem.