web: Add OverviewLogPane, a fork of LogPane
LogPane is pretty closely coupled with window-based scrolling and the particular
DOM layout of the old view, so probably we were going to need to fork it either way.

Meanwhile, I'm pretty unhappy with both the performance and flexibility of the old log pane.
It doesn't really fit in well with React's view of the world. React is optimized
so that any component can be live updated. But most logs need to be rendered once and
never touched again.

OverviewLogPane uses a lot of the existing CSS from LogPane, but substantially
changes rendering to vanilla JS, and gets rid of the concept of a render window.