tiltfile: add env support for local + local_resource (#4072)
tiltfile: add env support for local + local_resource

New, optional argument(s) for `local` and `local_resource` to pass
environment variables to the invoked command. In all cases, the
argument is a `dict` and must have both string keys and values:
no type conversion will take place.

For `local`, there is a single `env` argument.
For `local_resource`, there is `env`, which is used for the update
command (`cmd`) and `serve_env`, which is used for the serve command
(`serve_cmd`).

Any environment variables that are passed will take precedence over
the parent process' (i.e. `tilt`) environment. However, it's possible
to read from the environment in `Tiltfile`, so in the event that a
default is desired, a practical approach is to combine with `os.getenv`:

```
local_resource('mycmd', './mycmd', env={'KEY': os.getenv('KEY', 'default')})
```

Closes #3767.