engine: add initial touch points for local readiness probe
Currently a noop in that the readiness probe status is simply
initialized to successful and then never touched. In the future,
this will continue to be the behavior for local resources without
a readiness probe defined, while local resources with a readiness
probe will start off in a failing state until the process is up &
probe passes. (This aligns with the behavior of readiness probes
in K8s.)