store: make local resource HasEverBeenReadyOrSucceeded probe aware
Track the `LastReadyTime` similar to other resource types and gate
the return on this being non-zero. Local resources without a
readiness probe get this value initialized to now when the state
is first created (in theory, this could instead be updated when the
process starts, but the distinction is not particularly useful).