engine: handle readiness probes for local resources (#4102)
Local controller creates a probe worker when starting a serve command
if a probe spec is defined.

The worker will emit readiness probe status update actions that will
result in updates to the local runtime state, which internally tracks
the current passing state of the readiness probe. (Local resources
without a readiness probe will consider themselves to always have a
passing probe, but this is never exposed.)

Some adjustments to the way that "ready OR succeeded" is
tracked/computed align this closer to the K8s runtime state and try
to reduce the amount of logic going on inside of the state object.
This ensures that dependent resources won't start until the probe
is passing (if defined).