integration: add coverage for local readiness probes (#4130)
Expands the local resource test to cover readiness probes as well
as a resource dependency.

This also adds `[readiness probe]` as a prefix on each line from
the probe. The prepending is done in a buffer so that the output
can be written to the log in a single call to ensure that it's
consecutive and not broken up by any simultaneous resource logs.