web: clear logs functionality
New button in the resource bar to clear logs. On the all view,
this will clear logs for all resources. On an individual resource,
it will just clear that resource's logs.

If a resource currently has error/warnings, the latest build or
runtime span respectively will be retained. This is to avoid having
a bunch of alerts that just show empty text on click.