engine: do not start local resources with invalid probe specs (#4158)
Original behavior tried to fail gracefully, i.e. let the resource
start and just mark it as ready with a warning logged that the probe
spec was bad. There was a bug, however, which meant that dependent
resources would still see it as not ready and not start.

In the end, it seems that this behavior was more confusing than
helpful, so it's better to just prevent the resource from startup
if the probe spec is invalid.

Some validation has also been pushed down to the `Tiltfile` level,
which will provide better errors with line numbers. It's worth
noting, however, that there will likely always be some validation/
cases that we cannot catch during `Tiltfile` parsing, so the logic
in the controller is still important.