web: clear logs functionality (#4170)
New button in the resource bar to clear logs. On the all view,
this will clear logs for all resources. On an individual resource,
it will just clear that resource's logs.

Alerts for non-existent spans (i.e. because they've been removed)
will be ignored to prevent showing counts on the filters that won't
actually have any content.