server: make DI practical for controllers (#4250)
This is a bit roundabout, but the key issue is that controllers
(`Reconciler`s) need the controller client to do their work, which
comes from the `ctrl.Manager`, but isn't available until the manager
is actually connected to the API server, so it's not suitable for
normal DI.

A `ControllerBuilder` provides the glue between the manager and
reconcilers by providing them with the client and calling the
`SetupWithManager` methods on them. The latter of these is a standard
method that's bootstrapped by the codegen (but not part of the
`Reconciler` interface).

This also changes store subscribes to be torn down in the reverse
order that they're subscribed so dependencies can be handled; otherwise,
the API/HUD server will block forever shutting down since the controller
manager has active connections to it.