tests: fix engine dump on test failure
Randomized port assignment wasn't playing nice with this. I
simplified the port assignment so that it's part of the main Tilt
driver fixture so that all calls will use it consistently. In the
event that a test does pass `--port` as an arg, it won't include
`TILT_PORT` on the env for that call to avoid precedence issues.

Now, when `tilt dump engine` is called, it'll have the right port
via `TILT_PORT` in env, so that the engine state can be dumped.