tests: fix engine dump on test failure (#4324)
Randomized port assignment wasn't playing nice with this. I
simplified the port assignment so that it's part of the main Tilt
driver fixture so that all calls will use it consistently. In the
event that a test does pass `--port` as an arg or specifies
`TILT_PORT` in the env, it'll fail the test with an error that the
behaviors can't be mixed and matched.

Now, when `tilt dump engine` is called, it'll have the port
via the fixture, so that the engine state can be dumped
successfully rather than throwing an error about no Tilt
instance running.