tests: add explicit dump engine call to an integration test (#4342)
Currently, the integration tests (try to) invoke an engine dump
on failure to give state for debugging purposes.

We had a regression recently, however, where the engine dump
command was completely non-functional due to a serialization
issue. This adds an explicit call to dump the engine in one of
the tests which will catch any total breakage like this. It's
not a comprehensive test of the actual output of the engine
dump (though the CLI command ensures it's valid JSON), but
that's okay because the API of it is not stable and we're
moving away from it towards a stable interface for this type
of introspection via apiserver.