assets: report error if webpack dev port in use (#4344)
Webpack listens on IPv4, Go will happily bind to the same port
when we do an upfront check for the port being open.

In theory, we could just let Webpack error out and let us know
that the port is in use, but because it thinks it's running in
interactive mode, it helpfully pops up a message and offers to
let you pick a different port, which we'll never see. Running
it without any stdin attached _would_ cause it to error and
show the issue on macOS at least, but we started attaching a
dummy stdin in #3388 for other Webpack/CRA issues in the first
place (likely OS-specific), so this is a reasonable compromise.