tiltrun: add data model
The `TiltRun` type provides an aggregation of resources for a given
Tiltfile.

The goal is to fulfill the needs of the exit controller, which is
used in CI mode to either signal an exit with an error if a failure
occurs (e.g. failed build or runtime error) or clean exit if all
resources build + come up cleanly.