engine: consider Unschedulable pods as being in error state (#4386)
If a pod is unschedulable for some reason (e.g. resource limits),
mark its status as an error after a short (15 sec) grace period
instead of letting it sit in pending indefinitely.

If the scheduling issue is resolved, the resource (i.e. the pod
from Tilt's perspective) should then transition to OK as usual.

This makes it so that `tilt ci` will return an error if it can't schedule
pods due to a cluster issue rather than hanging indefinitely.

Closes #4385.