engine: use new apiserver Container model (#4445)
Because of how much K8s related code and pieces there are,
maintaining largely duplicative models is becoming cumbersome.

In trying to disentangle some of the pieces here, this will
start using the API model for a `Container`, which is almost
identical to the prior store version.

The biggest change is that the state representation now more
closely mirrors K8s and the other APIs. This is richer than
what the store model previously had, so there's some fairly
mechanical changes to look at whether fields are nil rather
than boolean.

Another minor change is that we can't store a `reference.Named`
on the API object, so the image as presented by K8s is stored
untouched. There is actually only one place that needs it,
which already conditionally filters out for a variety of reasons,
so it simply does the parsing and handles the error there (by
ignoring it, as was done previously). In practice, it should
never be able to fail parsing (or it wouldn't have been accepted
by K8s in the first place), so there should not really be a
difference here.

Lastly, the type definitions around `string` for container ID
and name are gone; these are now just strings in the API. The
engine converts them back to strongly-typed helpers upon
consumption for its own internal use. (These are useful within
internal code, but not really canonical in the apiserver models
at the moment. If desired, they can trivially be reintroduced
there in the future.)

Note that to limit the blast radius somewhat, a type alias is
used transitionally. A separate commit that eliminates that and
updates imports throughout the codebase will be done as its own
PR to avoid making this commit (which includes a handful of actual
logic changes) unreadable.