engine: use new apiserver Pod model (#4448)
See beb957c5cb45f24e649614c97b39116225ac11ed for similar changes
for Container.

The TL;DR is that this is being done as a precursor to moving more
K8s logic (Pod discovery) to apiserver to limit the number of almost
the same but slightly different data models being converted and passed
around.

The apiserver Pod data model is almost identical to the store Pod
model aside from renaming a couple fields to better align with the
K8s API names, so there aren't really any logic changes. It is already
simplified from the K8s model in a couple ways (aggregated status
messages & kubectl-esque status) that don't compromise the richness
of the data, so are left as-is besides renaming.