engine: use v1alpha.Pod for pod change events
This is another incremental change; it _mostly_ shifts the conversion
logic out of the action handler. The other biggest change is that it
now completely replaces the tracked Pod objects instead of getting a
reference and mutating it - this shifts some of the logic about
whether to track up a level. This gets the logic closer to its
eventual state as part of a reconciler based on the patterns we've
been following.

The actual action is still a change for a single Pod rather than a
`PodDiscoveryStatus` update, but the changes are large enough for
this chunk that it made since to use this as an increment.