cli: fix logs follow after server truncation
The client was using server checkpoint values to determine which
logs to print. These values aren't portable across `LogStore`s due
to truncation: internally an offset is used to find the right index
as the checkpoint value is monotonically increasing.

Instead, we can use the server value purely as a watermark for
filtering out duplicates, but use local checkpoints to determine
what logs to print / have been printed.