k8s: use KubernetesDiscovery spec to drive Pod discovery
This is another incremental change (but the largest) towards moving
Pod discovery fully into apiserver.

This adds `KubernetesDiscoveryXxxAction` action types for CRUD and
corresponding reducers that just update store (no apiserver
interaction yet).

Then, `k8swatch.ManifestSubscriber` is a new store subscriber that
takes on the job of creating `KubernetesDiscovery` specs for K8s
manifests. The existing `k8swatch.PodWatcher` subscriber now _only_
reacts to changes on `KubernetesDiscovery` objects and behaves like
a reconciler for them.

However, the status for `KubernetesDiscovery` is still empty and
no `KubernetesDiscoveryUpdateStatusAction` exists/is dispatched.
The existant `PodChangeAction` and its reducer are still used
downstream. This was just done to limit scope of this commit +
for code review sanity, and will be converted imminently.

Internally to `k8swatch.PodWatcher`, there were a few main chunks of
logic changes:
 * Diff logic acts on a per-entity basis rather than global and
   does not use the shared `watcherKnownState` helper anymore
 * Key everything off of `types.NamespacedName` and only map back
   to manifest for `PodChangeAction` dispatch as a bridge
 * Extra state tracking due to reconciling on a per-entity basis
   to ensure that namespace watches have appropriate lifecycle