engine: store object reference instead of just UID (#4465)
This is in preparation for `PodDiscovery` API work.

The motivation here is that currently, the engine extracts some
details about the objects it's managing from the specs, i.e. what
it "expects" to deploy. These are gathered across all manifests to
determine which namespaces to watch in the cluster.

Indepedently, the deployer stores the UIDs, but that loses the
mapping back to namespace. Currently, it's not an issue because of
the aforementioned logic determining which namespaces to watch.

With the move to apiserver, the semantics will be slightly different;
each manifest will define the ancestor UID + namespaces it cares
about. The reconciler will handle ensuring only 1x watch per namespace
is created for efficiency, but the key is that the spec needs to
include the UID + namespace together.

By holding the deployed object refs, we now have this information, so
a targeted spec can be created.

(This will also enable resolving a long-standing bug/unexpected
behavior that the extra label selectors will be across _any_ namespace
that's being watched, even if the respective manifest doesn't deploy
anything there. It's a very minor issue, but will be nice to clean up
in the process for free.)