api: helpers to avoid time granularity comparison issues (#4458)
From K8s, there are two time types for apiserver: `metav1.Time` and
`metav1.MicroTime`. The former has second-level granularity when
serialized and the latter microsecond-level granularity when
serialized.

Internally, these are both wrappers around Go stdlib `time.Time` and
only truncate when serialized. This has caused several issues due to
holding onto unserialized versions that are later compared against a
deserialized version resulting in false positives.

This was already largely mitigated by changes in the test fakes to
detect this and return an error. However, as part of migration work,
we often have apiserver entities that _don't_ live in the apiserver
for a period of time, so this doesn't detect that. New constructors
that perform the truncation immediately can help ensure that the
representation is the same upfront as it will be post-serialization.

There is another class of bugs that remains, however, which is that
we often have stdlib times for tracking purposes that we then compare
against apiserver times, and the difference in granularity (precision)
can cause bugs. To help avoid this, a new package, `timecmp`, offers
methods that will normalize the time values to the lowest-precision
of the two before doing a comparison. This has a side benefit that
even when comparing two stdlib `time.Time` values, it strips off the
monotonic clock portion. Because this is never present on Windows
(and possibly other platforms), there are often subtle
platform-specific bugs that this should find sooner.