k8s: move UID "claim" logic to ManifestSubscriber
The `PodWatcher` _mostly_ no longer cares if > 1 spec is watching
a Pod...assuming it's by UID (direct or transitive). Extra label
selectors are still ONLY applied if NO spec matches by UID and even
then only the FIRST spec to match gets the event.

This is all to prevent duplicate events for the same Pod across
manifests which causes all kinds of weird behavior elsewhere in
the engine at the moment.