engine: add multi-manifest support for Pods (#4498)
99.9% of the time, a Pod will only be associated with one manifest,
both because that's the general usage pattern aside from some Helm
charts that cause the same Pod to get referenced across mulitple
manifests, and because the `k8swatch.ManifestSubscriber` that is
responsible for creating `KubernetesDiscovery` objects from the
resources in the `Tiltfile` only allows one manifest to "claim" a
given UID so that even if it's referenced in another, it will be
invisible to it.

However, there is a remaining scenario, which is trickier, and that
is extra pod label selectors. The manifest subscriber can't
practically play the same games with uniqueness here (at least not
in any sane, predictable fashion). The `PodWatcher` now behaves
more like a reconciler (and will be one soon), so tries to evaluate
things on a per-entity basis rather than a global basis, which is
to say that without a lot of extra tracking to provide a similar
"claim" system on labels, it soon won't be able to guarantee that
only ONE `KubernetesDiscovery` object sees a particular Pod based
on a label match (and even then if and only if no other object had
a transitive UID match for that Pod!)

As a result, the path of least resistance is to ensure that Tilt
generally behaves in the event that the same pod occurs in multiple
resources. The manifest subscriber that generates the specs will
still try to avoid this by claiming UIDs, but in the case that extra
label selectors are being used and overlap with known UIDs, the pod
will show up in both resources without causing things to get wonky.