test: emit Pod events instead of manually dispatching actions (#4513)
The upper/buildcontroller tests (which share a fixture) were
simulating pod events by manually dispatching events, thus bypassing
the `PodWatcher` entirely.

That in and of itself isn't hugely problematic (although since these
are more integration than unit tests, it was skipping a lot of logic
that's useful to exercise).

The problem arises from the fact that events are dispatched per-pod
(`PodChangeAction`) currently, but as part of the changes for api
server, the `KubernetesDiscovery` status includes _all_ pods. Since
thigns are dispatched async, this presents new synchronization to
manage in the test.

Instead of trying to work around it further, this changes the tests
to rely on emitting Pod events from the fake K8s client and letting
`PodWatcher` see them and dispatch events.

This does present a complication because the pods generated by the
`PodWatcher` don't have any real relationship to the YAML in the
manifest. A new map on the fake build and deployer uses the provided
entities instead of trying to derive things from the YAML if
populated to accommodate this.