test: use PodWatcher in rest of upper_test tests (#4515)
See #4513 for the first big chunk of this. These were remaining
tests that used some different helper methods or dispatched stuff
directly themselves so made sense to split independently.

The only material change to a test here is for `TestPodEventOrdering`
which uses some newly exposed options on `PodBuilder` (from #4514)
to more accurately represent the data -- specifically, it ensures
that an old Pod looks like it would in reality: it has the same
name but a different UID. The latter part was missing before and is
critical to ensure the `PodWatcher` handles it correctly in terms
of linking to ancestor.