api: add MonitorStartTime field to KubernetesDiscovery status
Mirroring the field name/pattern in `FileWatchStatus`. This can
be generally useful as a sanity check that things are working, but
it's mostly useful now for tests to use for synchronization, so
that they can create a manifest and wait for the spec to appear
AND be active before continuing.

(NOTE: Nothing populates this field yet.)