podlogstream: use KubernetesDiscovery for changes
The `Pods` changes no longer exists since `PodChangeAction` is gone.
This meant that no `PodLogStream`s were ever getting created since
it appeared nothing was ever changing.

In the future, this Store subscriber will go way entirely and instead
the `KubernetesDiscovery` reconciler will create `PodLogStream` objects
directly, so despite it not currently really using the KD object, it
simulates the same flow (and avoids it being run excessively on
unrelated changes to help perf slightly). Note that it was not actually
using the Pod info previously either, which is a good thing since it
doesn't have access to it anymore!